% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem2mrgsolve.R
\name{nonmem2mrgsolve}
\alias{nonmem2mrgsolve}
\title{Convert NONMEM run into mrgsolve code}
\usage{
nonmem2mrgsolve(
  filename = NULL,
  dir = NULL,
  sigdig = NULL,
  write = T,
  return.orig = F
)
}
\arguments{
\item{filename}{String of the NONMEM model file name without any extension}

\item{dir}{String of the directory path to the NONMEM run files}

\item{sigdig}{Numeric of the number of significant digits to round non-fixed thetas and etas to; default NULL for no rounding}

\item{write}{Logical for whether to output the originally read in NONMEM ctl and ext files (\code{T} or \code{F})}
}
\value{
R dataframe of the mrgsolve code
}
\description{
Translates a NONMEM run into mrgsolve code format using the NONMEM ctl and ext files.
}
\examples{
nonmem2mrgsolve()

}
