% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemrgsolve.R
\name{writemrgsolve}
\alias{writemrgsolve}
\title{Write the mrgsolve code to an R file}
\usage{
writemrgsolve(mrg_code = NULL, filename = "mrgsolve_code0", dir = NULL)
}
\arguments{
\item{mrg_code}{Dataframe of the NONMEM model translated into mrgsolve code}

\item{filename}{String of the name for the mrgsolve output file without the .R extension}

\item{dir}{String of the directory path to the NONMEM run files}
}
\value{
R file of the mrgsolve code
}
\description{
Writes the mrgsolve code, translated from the input NONMEM run, to an R file.
}
\examples{
# writemrgsolve()

}
