% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemrgsolve.R
\name{writemrgsolve}
\alias{writemrgsolve}
\title{Write the mrgsolve code to a R file}
\usage{
writemrgsolve(mrg_code = NULL, filename = "mrgsolve_code0", dir = NULL)
}
\arguments{
\item{mrg_code}{Dataframe of the NONMEM run translated into mrgsolve code format}

\item{filename}{String of the file name for the output mrgsolve code R file}

\item{dir}{String of the directory path to the NONMEM run files}
}
\value{
R file of the mrgsolve code
}
\description{
Writes the mrgsolve code, translated from the input NONMEM run, to a R file.
}
\examples{
writemrgsolve()

}
